Added gdt support

+++ types.h
@@ -6,7 +6,7 @@ typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef int int32_t;
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long int uint64_t;
 typedef unsigned int size_t;
 
 #endif
